Refactored private utilities (#1837)
* refactor: Removed unused code

* refactor: Removed EmptyObject

* chore: Added changeset

---------
