+++ .changeset/rotten-olives-flash.md
@@ -1,0 +1,5 @@
+---
+"ember-intl": patch
+---
+
+Refactored private utilities
