+++ ember-intl/addon/types.d.ts
@@ -1,10 +1,10 @@
 import type { Formats as BaseFormats } from 'intl-messageformat';
 
 import type { RelativeTimeFormatOptions } from './-private/formatters/format-relative';
-import type { NestedStructure } from './-private/utils/flatten';
+import type { Translations } from './-private/utils/translations';
 
 export interface Formats extends Partial<BaseFormats> {
   relative?: Record<string, RelativeTimeFormatOptions>;
 }
 
-export type Translations = NestedStructure<string | number>;
+export type { Translations };
